refactor(#491): move getMaxWineYield to wineService
