+++ src/services/cellar.js
@@ -5,12 +5,9 @@
 
 import { v4 as uuid } from 'uuid'
 
-import { fermentableItemsMap, itemsMap } from '../data/maps'
+import { fermentableItemsMap } from '../data/maps'
 import { memoize } from '../utils/memoize'
-import { getInventoryQuantityMap } from '../utils/getInventoryQuantityMap'
 import { getYeastRequiredForWine } from '../utils/getYeastRequiredForWine'
-// eslint-disable-next-line no-unused-vars
-import { grapeVariety as grapeVarietyEnum } from '../enums'
 
 import { PURCHASEABLE_CELLARS } from '../constants'
 
@@ -59,41 +56,6 @@ export class CellarService {
       cellarInventory.length <
       (PURCHASEABLE_CELLARS.get(purchasedCellar)?.space ?? 0)
     )
-  }
-
-  // FIXME: Move this to WineService
-  // FIXME: Test this
-  /**
-   * @param {item} grape
-   * @param {{ id: string, quantity: number }[]} inventory
-   * @param {Array.<keg>} cellarInventory
-   * @param {number} cellarSize
-   * @param {grapeVarietyEnum} grapeVariety
-   * @returns {number}
-   */
-  getMaxWineYield = (
-    grape,
-    inventory,
-    cellarInventory,
-    cellarSize,
-    grapeVariety
-  ) => {
-    const {
-      [grape.id]: grapeQuantityInInventory = 0,
-      [itemsMap.yeast.id]: yeastQuantityInInventory = 0,
-    } = getInventoryQuantityMap(inventory)
-
-    const maxWineYieldPotential = Math.floor(
-      yeastQuantityInInventory / getYeastRequiredForWine(grapeVariety)
-    )
-
-    const maxYield = Math.min(
-      cellarSize - cellarInventory.length,
-      grapeQuantityInInventory,
-      maxWineYieldPotential
-    )
-
-    return maxYield
   }
 }
