+++ src/services/wine.js
@@ -10,6 +10,8 @@ import { WINE_INTEREST_RATE, WINE_GROWTH_TIMELINE_CAP } from '../constants'
 import { wineVarietyValueMap } from '../data/crops/grape'
 import { itemsMap } from '../data/maps'
 import { recipeType } from '../enums'
+import { getInventoryQuantityMap } from '../utils/getInventoryQuantityMap'
+import { getYeastRequiredForWine } from '../utils/getYeastRequiredForWine'
 
 export class WineService {
   /**
@@ -67,6 +69,40 @@ export class WineService {
    */
   isWineRecipe = recipe => {
     return 'recipeType' in recipe && recipe.recipeType === recipeType.WINE
+  }
+
+  // FIXME: Test this
+  /**
+   * @param {item} grape
+   * @param {{ id: string, quantity: number }[]} inventory
+   * @param {Array.<keg>} cellarInventory
+   * @param {number} cellarSize
+   * @param {grapeVarietyEnum} grapeVariety
+   * @returns {number}
+   */
+  getMaxWineYield = (
+    grape,
+    inventory,
+    cellarInventory,
+    cellarSize,
+    grapeVariety
+  ) => {
+    const {
+      [grape.id]: grapeQuantityInInventory = 0,
+      [itemsMap.yeast.id]: yeastQuantityInInventory = 0,
+    } = getInventoryQuantityMap(inventory)
+
+    const maxWineYieldPotential = Math.floor(
+      yeastQuantityInInventory / getYeastRequiredForWine(grapeVariety)
+    )
+
+    const maxYield = Math.min(
+      cellarSize - cellarInventory.length,
+      grapeQuantityInInventory,
+      maxWineYieldPotential
+    )
+
+    return maxYield
   }
 }
