+++ src/utils/getYeastRequiredForWine.js
@@ -1,4 +1,5 @@
 import { wineVarietyValueMap } from '../data/crops/grape'
+import { YEAST_REQUIREMENT_MULTIPLIER } from '../constants'
 // eslint-disable-next-line no-unused-vars
 import { grapeVariety as grapeVarietyEnum } from '../enums'
 
@@ -6,6 +7,5 @@ import { grapeVariety as grapeVarietyEnum } from '../enums'
  * @param {grapeVarietyEnum} grapeVariety
  */
 export const getYeastRequiredForWine = grapeVariety => {
-  // FIXME: Make the number a constant
-  return wineVarietyValueMap[grapeVariety] * 5
+  return wineVarietyValueMap[grapeVariety] * YEAST_REQUIREMENT_MULTIPLIER
 }
